# SwiftShortCourse/document/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.2",
        "title": "iFarm",
        "description": "Restful API List used for communication between Gateway/Mobile and Server",
        "contact": {
            "name": "Techmaster team",
            "email": "cuong@techmaster.vn",
            "url": "https://techmaster.vn"
        }
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
...
        "/refresh-token": {
            "post": {
                "description": "Use for Prolong Access Token",
                "operationId": "refreshToken",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "refreshToken",
                        "description": "Refresh Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "accessToken": {
                                    "type": "string"
                                },
                                "expiredAt": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "refreshToken": {
                                    "type": "string"
                                },
                                "authScheme": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Error": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}